perf: Only apply default settings when user is created or settings are requested
