+++ lib/Listener/UserCreatedListener.php
@@ -8,7 +8,6 @@ declare(strict_types=1);
 
 namespace OCA\Notifications\Listener;
 
-use OCA\Notifications\AppInfo\Application;
 use OCA\Notifications\Model\Settings;
 use OCA\Notifications\Model\SettingsMapper;
 use OCP\EventDispatcher\Event;
@@ -35,20 +34,7 @@ class UserCreatedListener implements IEventListener {
 
 		$userId = $event->getUser()->getUID();
 
-		$defaultSoundNotification = $this->config->getAppValue(Application::APP_ID, 'sound_notification') === 'yes' ? 'yes' : 'no';
-		$defaultSoundTalk = $this->config->getAppValue(Application::APP_ID, 'sound_talk') === 'yes' ? 'yes' : 'no';
-		$defaultBatchtime = (int)$this->config->getAppValue(Application::APP_ID, 'setting_batchtime');
-
-		if ($defaultBatchtime !== Settings::EMAIL_SEND_WEEKLY
-			&& $defaultBatchtime !== Settings::EMAIL_SEND_DAILY
-			&& $defaultBatchtime !== Settings::EMAIL_SEND_3HOURLY
-			&& $defaultBatchtime !== Settings::EMAIL_SEND_HOURLY
-			&& $defaultBatchtime !== Settings::EMAIL_SEND_OFF) {
-			$defaultBatchtime = Settings::EMAIL_SEND_3HOURLY;
-		}
-
-		$this->config->setUserValue($userId, Application::APP_ID, 'sound_notification', $defaultSoundNotification);
-		$this->config->setUserValue($userId, Application::APP_ID, 'sound_talk', $defaultSoundTalk);
-		$this->settingsMapper->setBatchSettingForUser($userId, $defaultBatchtime);
+		// Initializes the default settings
+		$this->settingsMapper->getSettingsByUser($userId);
 	}
 }
